refactor: 폴더명 임시 변경 #439
