+++ frontend/src/components/layout/Header/GameHeader/GameHeader.tsx
@@ -3,7 +3,7 @@ import { useParams } from 'react-router-dom';
 
 import { gameTitle, headerLayout, roundText } from '../Header.styled';
 
-import A11yOnly from '@/components/common/A11yOnly/A11yOnly';
+import A11yOnly from '@/components/common/A11yOnly2/A11yOnly';
 import { convertMsecToSecond } from '@/pages/GamePage/components/SelectContainer/Timer/Timer.util';
 
 import { useBalanceContentQuery, useFocus } from '@/hooks';
